backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.2.1
cryptography==46.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.2.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
flask-restx==1.3.0
